Seeder

+++ database/migrations/2022_09_26_102243_create_seasons_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('seasons', function (Blueprint $table) {
             $table->id();
-            $table->tinyint('nr');
+            $table->integer('nr');
             $table->string('name');
             $table->timestamps();
         });
